Refac : seperate recommend-list component
