+++ src/components/main/recent-style.tsx
@@ -10,9 +10,7 @@ import { apiGet } from "../../utils/request";
 import MainProduct from "./product-item";
 
 const RecentStyle: NextPage = () => {
-  const { data: products } = useQuery("products", apiGet.GET_ITEMS, {
-    suspense: true,
-  });
+  const { data: products } = useQuery("products", apiGet.GET_ITEMS);
   const [recentItems, setRecentItems] = useState<ProductData[]>([]);
 
   const { next, prev, transitionEnd, translateX, isMoving } = useSlide({
@@ -21,9 +19,10 @@ const RecentStyle: NextPage = () => {
   });
 
   useEffect(() => {
+    if (!products) return;
     const recents = products.slice(-10);
     setRecentItems(recents.reverse());
-  }, []);
+  }, [products]);
 
   return (
     <div className="px-5">
